feat: authorize routes with token

+++ server/Controllers/EventsController.cs
@@ -1,12 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using server.Context;
 using server.Models;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using server.Services;
 
 namespace server.Controllers
 {
@@ -15,32 +13,50 @@ namespace server.Controllers
     public class EventsController : ControllerBase
     {
         private readonly ApplicationContext _context;
+        private readonly IUserService _userService;
 
-        public EventsController(ApplicationContext context)
+        public EventsController(ApplicationContext context, IUserService userService)
         {
             _context = context;
+            _userService = userService;
         }
 
         // GET: api/Events
         [HttpGet]
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         public async Task<ActionResult<IEnumerable<Event>>> GetEvents()
         {
             if (_context.Events == null)
             {
                 return NotFound();
             }
-            return await _context.Events.ToListAsync();
+
+            Guid userId = _userService.GetUserId();
+
+            var events = await _context.Events
+                .Where(e => e.UserId == userId)
+                .ToListAsync();
+
+            return events;
         }
 
         // GET: api/Events/5
         [HttpGet("{id}")]
+        [Authorize]
         public async Task<ActionResult<Event>> GetEvent(Guid id)
         {
             if (_context.Events == null)
             {
                 return NotFound();
             }
+
+            if (!IsAuthorized(id))
+            {
+                return Unauthorized();
+            }
+
             var @event = await _context.Events.FindAsync(id);
+
 
             if (@event == null)
             {
@@ -53,6 +69,7 @@ namespace server.Controllers
         // PUT: api/Events/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
+        [Authorize]
         public async Task<IActionResult> PutEvent(Guid id, Event @event)
         {
             if (id != @event.Id)
@@ -61,6 +78,12 @@ namespace server.Controllers
             }
 
             _context.Entry(@event).State = EntityState.Modified;
+
+
+            if (!IsAuthorized(id))
+            {
+                return Unauthorized();
+            }
 
             try
             {
@@ -84,6 +107,7 @@ namespace server.Controllers
         // POST: api/Events
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
+        [Authorize]
         public async Task<ActionResult<Event>> PostEvent(Event @event)
         {
             if (_context.Events == null)
@@ -91,7 +115,7 @@ namespace server.Controllers
                 return Problem("Entity set 'ApplicationContext.Events'  is null.");
             }
 
-            if (@event.Title == null || @event.StartDate == DateTime.MinValue || @event.UserId == Guid.Empty)
+            if (string.IsNullOrEmpty(@event.Title) || @event.StartDate == DateTime.MinValue || @event.UserId == Guid.Empty)
             {
                 return BadRequest("Title, StartTime, and UserId are required.");
             }
@@ -113,6 +137,7 @@ namespace server.Controllers
 
         // DELETE: api/Events/5
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<IActionResult> DeleteEvent(Guid id)
         {
             if (_context.Events == null)
@@ -120,9 +145,15 @@ namespace server.Controllers
                 return NotFound();
             }
             var @event = await _context.Events.FindAsync(id);
+
             if (@event == null)
             {
                 return NotFound();
+            }
+
+            if (!IsAuthorized(id))
+            {
+                return Unauthorized();
             }
 
             _context.Events.Remove(@event);
@@ -134,6 +165,11 @@ namespace server.Controllers
         private bool EventExists(Guid id)
         {
             return (_context.Events?.Any(e => e.Id == id)).GetValueOrDefault();
+        }
+        private bool IsAuthorized(Guid eventId)
+        {
+            Guid userId = _userService.GetUserId();
+            return _context.Events.Any(e => e.Id == eventId && e.UserId == userId);
         }
     }
 }
